Remove temporary or autocreated workspace/timeline files from the playground
