+++ Diffy.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:Diff.swift.xcodeproj">
+      location = "self:Diffy.xcodeproj">
    </FileRef>
 </Workspace>
