+++ Graph.playground/Sources/ExtendedDiff.swift
@@ -1,1 +1,1 @@
-../../Sources/Diff/ExtendedDiff.swift
+../../Sources/Diffy/ExtendedDiff.swift
