+++ Graph.playground/Sources/Patch.swift
@@ -1,1 +1,1 @@
-../../Sources/Diff/Patch.swift
+../../Sources/Diffy/Patch.swift
