--- Graph.playground/timeline.xctimeline
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Timeline
-   version = "3.0">
-   <TimelineItems>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/tonyarnold/Documents/Repositories/Diff.swift/Graph.playground#CharacterRangeLen=0&amp;CharacterRangeLoc=163&amp;EndingColumnNumber=70&amp;EndingLineNumber=5&amp;StartingColumnNumber=70&amp;StartingLineNumber=5&amp;Timestamp=527080677.43884"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/tonyarnold/Documents/Repositories/Diff.swift/Graph.playground#CharacterRangeLen=0&amp;CharacterRangeLoc=163&amp;EndingColumnNumber=70&amp;EndingLineNumber=5&amp;StartingColumnNumber=70&amp;StartingLineNumber=5&amp;Timestamp=527080677.438998"
-         lockedSize = "{280, 50}"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/tonyarnold/Documents/Repositories/Diff.swift/Graph.playground#CharacterRangeLen=19&amp;CharacterRangeLoc=174&amp;EndingColumnNumber=20&amp;EndingLineNumber=6&amp;StartingColumnNumber=1&amp;StartingLineNumber=6&amp;Timestamp=527080677.439172"
-         lockedSize = "{1033, 752}"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-   </TimelineItems>
-</Timeline>
