+++ Sources/Diffy/GenericPatch.swift
@@ -1,4 +1,3 @@
-
 struct SortedPatchElement<T> {
     var value: Patch<T>
     let sourceIndex: Int
@@ -52,7 +51,7 @@ func shiftPatchElement<T>(from: DoublyLinkedList<SortedPatchElement<T>>, to: Dou
     switch type {
     case .cycle:
         fatalError()
-    case .neighbor(let direction), .jump(let direction):
+    case let .neighbor(direction), let .jump(direction):
         if case .left = direction {
             switch (from.value.value, to.value.value) {
             case (.insertion, _):
@@ -107,7 +106,7 @@ func shiftedPatchElements<T>(from sortedPatchElements: [SortedPatchElement<T>]) 
     }
 
     guard let result = linkedList?.array().sorted(by: { (fst, second) -> Bool in
-        return fst.sortedIndex < second.sortedIndex
+        fst.sortedIndex < second.sortedIndex
     }) else {
         return []
     }
