+++ Sources/Diffy/NestedDiff.swift
@@ -1,15 +1,12 @@
-
-public struct NestedExtendedDiff: DiffProtocol {
+public struct NestedDiff: DiffProtocol {
 
     public typealias Index = Int
 
     public enum Element {
         case deleteSection(Int)
         case insertSection(Int)
-        case moveSection(from: Int, to: Int)
         case deleteElement(Int, section: Int)
         case insertElement(Int, section: Int)
-        case moveElement(from: (item: Int, section: Int), to: (item: Int, section: Int))
     }
 
     /// Returns the position immediately after the given index.
@@ -24,8 +21,6 @@ public struct NestedExtendedDiff: DiffProtocol {
     public let elements: [Element]
 }
 
-public typealias NestedElementEqualityChecker<T: Collection> = (T.Iterator.Element.Iterator.Element, T.Iterator.Element.Iterator.Element) -> Bool where T.Iterator.Element: Collection
-
 public extension Collection
     where Iterator.Element: Collection {
 
@@ -33,31 +28,22 @@ public extension Collection
     ///
     /// - parameter other: a collection to compare the calee to
     /// - returns: a `NestedDiff` between the calee and `other` collection
-    public func nestedExtendedDiff(
+    public func nestedDiff(
         to: Self,
         isEqualSection: EqualityChecker<Self>,
         isEqualElement: NestedElementEqualityChecker<Self>
-    ) -> NestedExtendedDiff {
-
-        // FIXME: This implementation is a copy paste of NestedDiff with some adjustments.
-
+    ) -> NestedDiff {
         let diffTraces = outputDiffPathTraces(to: to, isEqual: isEqualSection)
 
-        let sectionDiff =
-            extendedDiff(
-                from: Diff(traces: diffTraces),
-                other: to,
-                isEqual: isEqualSection
-            ).map { element -> NestedExtendedDiff.Element in
-                switch element {
-                case let .delete(at):
-                    return .deleteSection(at)
-                case let .insert(at):
-                    return .insertSection(at)
-                case let .move(from, to):
-                    return .moveSection(from: from, to: to)
-                }
+        // Diff sections
+        let sectionDiff = Diff(traces: diffTraces).map { element -> NestedDiff.Element in
+            switch element {
+            case let .delete(at):
+                return .deleteSection(at)
+            case let .insert(at):
+                return .insertSection(at)
             }
+        }
 
         // Diff matching sections (moves, deletions, insertions)
         let filterMatchPoints = { (trace: Trace) -> Bool in
@@ -66,26 +52,6 @@ public extension Collection
             }
             return false
         }
-
-        let sectionMoves =
-            sectionDiff.flatMap { diffElement -> (Int, Int)? in
-                if case let .moveSection(from, to) = diffElement {
-                    return (from, to)
-                }
-                return nil
-            }.flatMap { move -> [NestedExtendedDiff.Element] in
-                return itemOnStartIndex(advancedBy: move.0).extendedDiff(to.itemOnStartIndex(advancedBy: move.1), isEqual: isEqualElement)
-                    .map { diffElement -> NestedExtendedDiff.Element in
-                        switch diffElement {
-                        case let .insert(at):
-                            return .insertElement(at, section: move.1)
-                        case let .delete(at):
-                            return .deleteElement(at, section: move.0)
-                        case let .move(from, to):
-                            return .moveElement(from: (from, move.0), to: (to, move.1))
-                        }
-                    }
-            }
 
         // offset & section
 
@@ -101,21 +67,19 @@ public extension Collection
         }
 
         let elementDiff = zip(zip(fromSections, toSections), matchingSectionTraces)
-            .flatMap { (args) -> [NestedExtendedDiff.Element] in
+            .flatMap { (args) -> [NestedDiff.Element] in
                 let (sections, trace) = args
-                return sections.0.extendedDiff(sections.1, isEqual: isEqualElement).map { diffElement -> NestedExtendedDiff.Element in
+                return sections.0.diff(sections.1, isEqual: isEqualElement).map { diffElement -> NestedDiff.Element in
                     switch diffElement {
                     case let .delete(at):
                         return .deleteElement(at, section: trace.from.x)
                     case let .insert(at):
                         return .insertElement(at, section: trace.from.y)
-                    case let .move(from, to):
-                        return .moveElement(from: (from, trace.from.x), to: (to, trace.from.y))
                     }
                 }
             }
 
-        return NestedExtendedDiff(elements: sectionDiff + sectionMoves + elementDiff)
+        return NestedDiff(elements: sectionDiff + elementDiff)
     }
 }
 
@@ -124,11 +88,11 @@ public extension Collection
     Iterator.Element.Iterator.Element: Equatable {
 
     /// - seealso: `nestedDiff(to:isEqualSection:isEqualElement:)`
-    public func nestedExtendedDiff(
+    public func nestedDiff(
         to: Self,
         isEqualSection: EqualityChecker<Self>
-    ) -> NestedExtendedDiff {
-        return nestedExtendedDiff(
+    ) -> NestedDiff {
+        return nestedDiff(
             to: to,
             isEqualSection: isEqualSection,
             isEqualElement: { $0 == $1 }
@@ -141,11 +105,11 @@ public extension Collection
     Iterator.Element: Equatable {
 
     /// - seealso: `nestedDiff(to:isEqualSection:isEqualElement:)`
-    public func nestedExtendedDiff(
+    public func nestedDiff(
         to: Self,
         isEqualElement: NestedElementEqualityChecker<Self>
-    ) -> NestedExtendedDiff {
-        return nestedExtendedDiff(
+    ) -> NestedDiff {
+        return nestedDiff(
             to: to,
             isEqualSection: { $0 == $1 },
             isEqualElement: isEqualElement
@@ -159,8 +123,8 @@ public extension Collection
     Iterator.Element.Iterator.Element: Equatable {
 
     /// - seealso: `nestedDiff(to:isEqualSection:isEqualElement:)`
-    public func nestedExtendedDiff(to: Self) -> NestedExtendedDiff {
-        return nestedExtendedDiff(
+    public func nestedDiff(to: Self) -> NestedDiff {
+        return nestedDiff(
             to: to,
             isEqualSection: { $0 == $1 },
             isEqualElement: { $0 == $1 }
@@ -168,7 +132,7 @@ public extension Collection
     }
 }
 
-extension NestedExtendedDiff.Element: CustomDebugStringConvertible {
+extension NestedDiff.Element: CustomDebugStringConvertible {
     public var debugDescription: String {
         switch self {
         case let .deleteElement(row, section):
@@ -179,10 +143,6 @@ extension NestedExtendedDiff.Element: CustomDebugStringConvertible {
             return "IE(\(row),\(section))"
         case let .insertSection(section):
             return "IS(\(section))"
-        case let .moveElement(from, to):
-            return "ME((\(from.item),\(from.section)),(\(to.item),\(to.section)))"
-        case let .moveSection(from, to):
-            return "MS(\(from),\(to))"
         }
     }
 }
