+++ Sources/PerformanceTester/main.swift
@@ -1,6 +1,3 @@
-//
-// Copyright © 2017 Tony Arnold. All rights reserved.
-
 public final class PerformanceTesterTool {
     private let arguments: [String]
 
@@ -13,23 +10,23 @@ public final class PerformanceTesterTool {
             throw Error.tooFewFilesSpecified
         }
 
-		let from = arguments[1]
-		let to = arguments[2]
+        let from = arguments[1]
+        let to = arguments[2]
 
-		print("Benchmarking Dwifft…", terminator: " ")
-		let dwifft = performDiff(fromFilePath: from, toFilePath: to, diffFunc: performDwifft)
-		print("Done!")
+        print("Benchmarking Dwifft…", terminator: " ")
+        let dwifft = performDiff(fromFilePath: from, toFilePath: to, diffFunc: performDwifft)
+        print("Done!")
 
-		print("Benchmarking Diff…", terminator: " ")
-		let diff = performDiff(fromFilePath: from, toFilePath: to, diffFunc: diffSwift)
-		print("Done!", terminator: "\n\n\n")
+        print("Benchmarking Diff…", terminator: " ")
+        let diff = performDiff(fromFilePath: from, toFilePath: to, diffFunc: diffSwift)
+        print("Done!", terminator: "\n\n\n")
 
-		print("         | Diff.swift          | Dwifft           ")
-		print("---------|:-------------------:|:----------------:")
-		print(" same    |   \(diff.same   )   | \(dwifft.same   )")
-		print(" created |   \(diff.created)   | \(dwifft.created)")
-		print(" deleted |   \(diff.deleted)   | \(dwifft.deleted)")
-		print(" diff    |   \(diff.changed)   | \(dwifft.changed)")
+        print("         | Diffy          | Dwifft           ")
+        print("---------|:-------------------:|:----------------:")
+        print(" same    |   \(diff.same)   | \(dwifft.same)")
+        print(" created |   \(diff.created)   | \(dwifft.created)")
+        print(" deleted |   \(diff.deleted)   | \(dwifft.deleted)")
+        print(" diff    |   \(diff.changed)   | \(dwifft.changed)")
     }
 }
