--- Supporting Files/Diff.h
@@ -1,19 +1,0 @@
-//
-//  Diff.swift.h
-//  Diff.swift
-//
-//  Created by Wojciech Czekalski on 26.04.2016.
-//  Copyright © 2016 wczekalski. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for Diff.swift.
-FOUNDATION_EXPORT double Diff_swiftVersionNumber;
-
-//! Project version string for Diff.swift.
-FOUNDATION_EXPORT const unsigned char Diff_swiftVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <Diff_swift/PublicHeader.h>
-
-
