+++ Supporting Files/Diffy.h
@@ -1,0 +1,7 @@
+@import Foundation;
+
+//! Project version number for Diffy.
+FOUNDATION_EXPORT double DiffyVersionNumber;
+
+//! Project version string for Diffy.
+FOUNDATION_EXPORT const unsigned char DiffyVersionString[];
