+++ Tests/DiffyTests/ExtendedPatchSortTests.swift
@@ -1,6 +1,5 @@
-
 import XCTest
-import Diff
+@testable import Diffy
 
 class ExtendedPatchSortTests: XCTestCase {
 
@@ -10,7 +9,7 @@ class ExtendedPatchSortTests: XCTestCase {
             ("Oh Hi", "Hi Oh", "M(0,4)M(0,4)M(0,2)"),
             ("12345", "12435", "M(2,3)"),
             ("1362", "31526", "M(0,2)M(1,3)I(2,5)"),
-            ("221", "122", "M(2,0)"),
+            ("221", "122", "M(2,0)")
         ]
 
         for expectation in expectations {
@@ -23,7 +22,7 @@ class ExtendedPatchSortTests: XCTestCase {
     func testInsertionDeletionMove() {
         let expectations = [
             ("gitten", "sitting", "I(5,i)I(1,s)D(5)M(0,6)"),
-            ("1362", "31526", "I(3,5)M(0,2)M(1,4)"),
+            ("1362", "31526", "I(3,5)M(0,2)M(1,4)")
         ]
 
         let sort: ExtendedSortingFunction = { fst, snd in
@@ -53,7 +52,7 @@ class ExtendedPatchSortTests: XCTestCase {
         let expectations = [
             ("gitten", "sitting", "D(4)M(0,4)I(0,s)I(4,i)"),
             ("1362", "31526", "M(0,2)M(1,3)I(2,5)"),
-            ("a1b2c3pq", "3sa1cz2rb", "D(7)D(6)M(5,0)M(3,5)M(3,4)I(1,s)I(5,z)I(7,r)"),
+            ("a1b2c3pq", "3sa1cz2rb", "D(7)D(6)M(5,0)M(3,5)M(3,4)I(1,s)I(5,z)I(7,r)")
         ]
 
         let sort: ExtendedSortingFunction = { fst, snd in
