+++ Tests/DiffyTests/PatchSortTests.swift
@@ -1,5 +1,5 @@
 import XCTest
-@testable import Diff
+@testable import Diffy
 
 class PatchTests: XCTestCase {
 
@@ -19,7 +19,7 @@ class PatchTests: XCTestCase {
             ("", "", ""),
             ("Oh Hi", "Hi Oh", "D(0)D(0)D(0)I(2, )I(3,O)I(4,h)"),
             ("1362", "31526", "D(0)D(1)I(1,1)I(2,5)I(4,6)"),
-            ("1234b2", "ab", "D(0)D(0)D(0)D(0)I(0,a)D(2)"),
+            ("1234b2", "ab", "D(0)D(0)D(0)D(0)I(0,a)D(2)")
         ]
 
         for expectation in defaultOrder {
@@ -44,18 +44,18 @@ class PatchTests: XCTestCase {
             ("1234", "1234", ""),
             ("", "", ""),
             ("Oh Hi", "Hi Oh", "I(5, )I(6,O)I(7,h)D(0)D(0)D(0)"),
-            ("1362", "31526", "I(3,1)I(4,5)I(6,6)D(0)D(1)"),
+            ("1362", "31526", "I(3,1)I(4,5)I(6,6)D(0)D(1)")
         ]
 
         let insertionsFirstSort = { (element1: Diff.Element, element2: Diff.Element) -> Bool in
             switch (element1, element2) {
-            case (.insert(let at1), .insert(let at2)):
+            case let (.insert(at1), .insert(at2)):
                 return at1 < at2
             case (.insert(_), .delete(_)):
                 return true
             case (.delete(_), .insert(_)):
                 return false
-            case (.delete(let at1), .delete(let at2)):
+            case let (.delete(at1), .delete(at2)):
                 return at1 < at2
             }
         }
@@ -85,18 +85,18 @@ class PatchTests: XCTestCase {
             ("1234", "1234", ""),
             ("", "", ""),
             ("Oh Hi", "Hi Oh", "D(0)D(0)D(0)I(2, )I(3,O)I(4,h)"),
-            ("1362", "31526", "D(0)D(1)I(1,1)I(2,5)I(4,6)"),
+            ("1362", "31526", "D(0)D(1)I(1,1)I(2,5)I(4,6)")
         ]
 
         let deletionsFirstSort = { (element1: Diff.Element, element2: Diff.Element) -> Bool in
             switch (element1, element2) {
-            case (.insert(let at1), .insert(let at2)):
+            case let (.insert(at1), .insert(at2)):
                 return at1 < at2
             case (.insert(_), .delete(_)):
                 return false
             case (.delete(_), .insert(_)):
                 return true
-            case (.delete(let at1), .delete(let at2)):
+            case let (.delete(at1), .delete(at2)):
                 return at1 < at2
             }
         }
@@ -113,8 +113,8 @@ class PatchTests: XCTestCase {
 
     func testRandomStringPermutationRandomPatchSort() {
 
-        let sort = { (element1: Diff.Element, element2: Diff.Element) -> Bool in
-            return arc4random_uniform(2) == 0
+        let sort = { (_: Diff.Element, _: Diff.Element) -> Bool in
+            arc4random_uniform(2) == 0
         }
         for _ in 0 ..< 200 {
             let randomString = randomAlphaNumericString(length: 30)
